use auto; s/atol/atoi/g

+++ Cxx11/transpose-opencl.cc
@@ -163,7 +163,7 @@ int main(int argc, char * argv[])
       }
 
       // order of a the matrix
-      order = std::atol(argv[2]);
+      order = std::atoi(argv[2]);
       if (order <= 0) {
         throw "ERROR: Matrix Order must be greater than 0";
       }
